// src/types/invidious.ts
export interface ThumbnailObject {
  quality: string;
  url: string;
  width: number;
  height: number;
}

export interface VideoObject {
  type: "video" | "shortVideo" | string;
  title: string;
  videoId: string;
  author: string;
  authorId: string;
  authorUrl: string;
  authorVerified?: boolean;
  videoThumbnails: ThumbnailObject[];
  authorThumbnails?: ThumbnailObject[];
  description?: string;
  descriptionHtml?: string;
  viewCount?: number;
  viewCountText?: string;
  lengthSeconds?: number;
  published?: number;
  publishedText?: string;
  premiereTimestamp?: number;
  liveNow?: boolean;
  premium?: boolean;
  paid?: boolean;
  isUpcoming?: boolean;
  isNew?: boolean;
  is4k?: boolean;
  is8k?: boolean;
  isVr180?: boolean;
  isVr360?: boolean;
  is3d?: boolean;
  hasCaptions?: boolean;
  playlistId?: string;
  playlistThumbnail?: string;
}
// ...
export interface AdaptiveFormat {
  index?: string;
  bitrate?: string;
  init?: string;
  url: string;
  itag?: string;
  type?: string;
  clen?: string;
  lmt?: string;
  projectionType?: string;
  fps?: number;
  container?: string;
  encoding?: string;
  qualityLabel?: string;
  resolution?: string;
}

export interface FormatStream {
  url: string;
  itag?: string;
  type?: string;
  quality?: string;
  bitrate?: string;
  container?: string;
  encoding?: string;
  qualityLabel?: string;
  resolution?: string;
  size?: string;
}

export interface Caption {
  label?: string;
  languageCode?: string;
  url?: string;
}

export interface MusicTrack {
  song?: string;
  album?: string;
  artist?: string;
  license?: string;
}

export interface RecommendedVideo extends VideoObject {
  recommended?: boolean;
}

export interface VideoDetails extends VideoObject {
  hlsUrl?: string;
  dashUrl?: string;
  adaptiveFormats: AdaptiveFormat[];
  formatStreams: FormatStream[];
  captions: Caption[];
  musicTracks: MusicTrack[];
  recommendedVideos: RecommendedVideo[];
  likeCount?: number;
  dislikeCount?: number;
  genre?: string;
  allowRatings?: boolean;
  subCount?: number;
  commentCount?: number;
}